Move setup_padding to util
